Improve type casts and add string format to syslog call
[git-p4: depot-paths = "//depot/ACG/Smith/Firmware/": change = 96100]


Former-commit-id: 5e3b5c0c35ec3efcb3a3dc11bbc48b43a356b77b

+++ C++/PrintData.cpp
@@ -224,5 +224,5 @@ int PrintData::gzOpenFrontend(char* pathname, int oflags, int mode)
             return -1;
     }
 
-    return (int)gzf;
+    return (intptr_t)gzf;
 }
